requirements.txt
attrdict==2.0.0
certifi==2018.4.16
chardet==3.0.4
cytoolz==0.9.0.1
Django==2.1
djangorestframework==3.8.2
eth-abi==1.1.1
eth-account==0.3.0
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
hexbytes==0.1.0
idna==2.7
lru-dict==1.1.6
parsimonious==0.8.0
pycryptodome==3.6.4
pypiwin32==223
pytz==2018.5
pywin32==223
requests==2.19.1
rlp==1.0.2
six==1.11.0
toolz==0.9.0
urllib3==1.23
web3==4.5.0
websockets==5.0.1
attrdict==2.0.0
certifi==2018.4.16
chardet==3.0.4
cytoolz==0.9.0.1
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
eth-abi==1.1.1
eth-account==0.3.0
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
hexbytes==0.1.0
idna==2.7
lru-dict==1.1.6
parsimonious==0.8.0
Pillow==5.2.0
pycryptodome==3.6.4
pypiwin32==223
pytz==2018.5
pywin32==223
requests==2.19.1
rlp==1.0.2
six==1.11.0
toolz==0.9.0
urllib3==1.23
web3==4.5.0
websockets==5.0.1
